+++ admin/partials/acc_user_importer-admin-settings.php
@@ -83,9 +83,21 @@ function accUM_verify_expiry_default()
 {
     return "off";
 }
+function accUM_get_delete_ex_users_default()
+{
+    return "off";
+}
 function accUM_get_ex_user_role_value_default()
 {
     return "subscriber";
+}
+function accUM_get_when_2_delete_ex_user_default()
+{
+    return 365;
+}
+function accUM_get_new_owner_default()
+{
+    return "";
 }
 function accUM_get_default_max_log_files()
 {
@@ -148,7 +160,19 @@ function accUM_get_verify_expiry()
     return $setting == "on";
 }
 
-// Returns true if we need to scan the DB looking for expired users
+// Returns true if we need to delete old expired users from database
+function accUM_get_delete_ex_users()
+{
+    $options = get_option("accUM_data");
+    if (!isset($options["accUM_delete_ex_users"])) {
+        $setting = accUM_get_delete_ex_users_default();
+    } else {
+        $setting = $options["accUM_delete_ex_users"];
+    }
+    return $setting == "on";
+}
+
+// Returns the configured list of users to synchronize
 function accUM_get_sync_list()
 {
     $options = get_option("accUM_data");
@@ -156,6 +180,30 @@ function accUM_get_sync_list()
         $setting = accUM_get_sync_list_default();
     } else {
         $setting = $options["accUM_sync_list"];
+    }
+    return $setting;
+}
+
+// Returns the number of days before deleting an expired user.
+function accUM_get_when_2_delete_ex_user()
+{
+    $options = get_option("accUM_data");
+    if (!isset($options["accUM_when_2_delete_ex_user"])) {
+        $setting = accUM_get_when_2_delete_ex_user_default();
+    } else {
+        $setting = $options["accUM_when_2_delete_ex_user"];
+    }
+    return $setting;
+}
+
+// Returns the new content owner when a user is deletec.
+function accUM_get_new_owner()
+{
+    $options = get_option("accUM_data");
+    if (!isset($options["accUM_new_owner"])) {
+        $setting = accUM_get_new_owner_default();
+    } else {
+        $setting = $options["accUM_new_owner"];
     }
     return $setting;
 }
@@ -196,6 +244,7 @@ function accUM_settings_init()
                 "BUGABOOS" => "BUGABOOS",
             ],
             "default" => accUM_get_section_default(),
+            "help" => "Select one",
         ]
     );
 
@@ -291,18 +340,6 @@ function accUM_settings_init()
     );
 
     add_settings_field(
-        "accUM_verify_expiry", //ID
-        "Also check user expiry in local DB",
-        "accUM_chkbox_render", //Callback
-        "acc_admin_page", //Page
-        "accUM_user_section", //Section
-        [
-            "name" => "accUM_verify_expiry",
-            "default" => accUM_verify_expiry_default(),
-        ]
-    );
-
-    add_settings_field(
         "accUM_new_user_role_action", //ID
         "When creating a new user, what should I do with role?",
         "accUM_select_render", //Callback
@@ -360,6 +397,67 @@ function accUM_settings_init()
             "name" => "accUM_ex_user_role_value",
             "values" => $roles,
             "default" => accUM_get_ex_user_role_value_default(),
+        ]
+    );
+
+    add_settings_field(
+        "accUM_verify_expiry", //ID
+        "Also check user expiry in local DB",
+        "accUM_chkbox_render", //Callback
+        "acc_admin_page", //Page
+        "accUM_user_section", //Section
+        [
+            "name" => "accUM_verify_expiry",
+            "default" => accUM_verify_expiry_default(),
+        ]
+    );
+
+    add_settings_field(
+        "accUM_delete_ex_users", //ID
+        "Delete expired user accounts after a while",
+        "accUM_chkbox_render", //Callback
+        "acc_admin_page", //Page
+        "accUM_user_section", //Section
+        [
+            "name" => "accUM_delete_ex_users",
+            "default" => accUM_get_delete_ex_users_default(),
+            "help" => "Requires 'Also check user expiry' option.",
+        ]
+    );
+
+    add_settings_field(
+        "accUM_when_2_delete_ex_user", //ID
+        "How many days before deleting expired users from database?",
+        "accUM_text_render", //Callback
+        "acc_admin_page", //Page
+        "accUM_user_section", //Section
+        [
+            "type" => "number",
+            "name" => "accUM_when_2_delete_ex_user",
+            "default" => accUM_get_when_2_delete_ex_user_default(),
+            "help" =>
+                "Enter the number of days after which to delete the user account.",
+        ]
+    );
+
+    add_settings_field(
+        "accUM_new_owner", //ID
+        "When deleting a user, who will become the new content owner?",
+        "accUM_text_render", //Callback
+        "acc_admin_page", //Page
+        "accUM_user_section", //Section
+        [
+            "type" => "text",
+            "name" => "accUM_new_owner",
+            "default" => accUM_get_new_owner_default(),
+            "help" =>
+                "Enter the new owner login name. Suggestion: manually " .
+                "create a dummy user (example: 'ex-member') to receive " .
+                "ownership of content for users we need to delete, " .
+                "and enter its login name here. The plugin will reassign " .
+                "posts, pages, articles, events. Leaving this box " .
+                "empty will delete the user content along with the user, " .
+                "and you might end up with missing pages or broken links.",
         ]
     );
 
@@ -457,7 +555,13 @@ function accUM_select_render($args)
         $select_value = $options[$input_name];
     }
 
-    $html = "<select id=\"$input_name\" name=\"accUM_data[$input_name]\">";
+    //if there is help text to display when hovering
+    $help = "";
+    if (!empty($args["help"])) {
+        $help = $args["help"];
+    }
+
+    $html = "<select id=\"$input_name\" name=\"accUM_data[$input_name] \" title=\"$help\">";
 
     //Fill columns
     if ($args["values"]) {
@@ -491,6 +595,13 @@ function accUM_chkbox_render($args)
     $html = "<input type=\"checkbox\"";
     $html .= " id=\"$input_name\"";
     $html .= " name=\"accUM_data[$input_name]\"";
+
+    //if there is help text to display when hovering
+    if (!empty($args["help"])) {
+        $help = $args["help"];
+        $html .= " title=\"$help\"";
+    }
+
     $html .= checked("on", $select_value, false) . " />";
     echo $html;
 }